The praxis read model moves out to services/praxis_out.py
Pure move, no behaviour change. `build_praxis_out`, `build_praxis_card_out`,
`build_praxis_cards`, `_build_member_out`, `_build_invite_out`,
`author_contributions_for`, `applied_metatasks_for`, `duel_id_map` and
`_resolve_card_scoring` leave `services/praxis.py` for a new read-model module.

The dependency runs one way only: `praxis_out` reads `services.praxis` for
`can_flag_praxis` (the domain predicate the `can_flag` response field reports)
and nothing reads back, so the two modules cannot form a cycle. That direction
is why consumers repoint rather than reading through a re-export shim: a shim
would make the domain service import its own read model.

Dissolved one function-body service import on the way: `build_praxis_out`'s
`from services.vote import viewer_can_vote` was function-scoped because
`services.vote` was a cycle risk from inside `services.praxis`. The read model
has no such edge, so it is module-level now.

Also dropped `from models.era import Era` from `services/praxis.py` and
`ModerationStatus` from `routers/praxes.py` — both were already dead on main.
