+++ backend/tests/integration/test_praxis_voted_filter.py
@@ -24,7 +24,8 @@ from models.faction import Faction
 from models.praxis import Praxis, PraxisMember, PraxisStatus, PraxisType
 from models.task import Task
 from models.vote import Vote
-from services.praxis import build_praxis_card_out, list_praxes, VotedFilter
+from services.praxis import list_praxes, VotedFilter
+from services.praxis_out import build_praxis_card_out
 from services.vote import cast_or_update_vote
 from services.vote_tally import viewer_votes_for
